refactor /visualizations routes into separate files
